Correcting the code after the first review
