+++ src/main/java/online/store/config/OpenApiConfig.java
@@ -1,0 +1,20 @@
+package online.store.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class OpenApiConfig {
+
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .openapi("3.0.1")
+                .info(new Info()
+                        .title("Online Store API")
+                        .version("1.0.0")
+                        .description("API для управления продуктами интернет-магазина"));
+    }
+}
