--- src/main/java/online/store/configs/JpaConfig.java
@@ -1,9 +1,0 @@
-package online.store.configs;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
