--- src/main/java/online/store/configs/OpenApiConfig.java
@@ -1,20 +1,0 @@
-package online.store.configs;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class OpenApiConfig {
-
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .openapi("3.0.1")
-                .info(new Info()
-                        .title("Online Store API")
-                        .version("1.0.0")
-                        .description("API для управления продуктами интернет-магазина"));
-    }
-}
