--- src/main/java/online/store/configs/SchedulerConfig.java
@@ -1,11 +1,0 @@
-package online.store.configs;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@Configuration
-@EnableScheduling
-@ConditionalOnProperty(name = "scheduler.enabled", matchIfMissing = true)
-public class SchedulerConfig {
-}
