+++ src/main/java/online/store/controller/impl/ExceptionApiControllerImpl.java
@@ -1,0 +1,60 @@
+package online.store.controller.impl;
+
+import online.store.controller.ExceptionApiController;
+import online.store.exception.ErrorMessage;
+import online.store.exception.ProductNotFoundException;
+import online.store.exception.UpdateProductException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+
+@ControllerAdvice
+public class ExceptionApiControllerImpl implements ExceptionApiController {
+
+
+    @Override
+    public ResponseEntity<ErrorMessage> notFoundException(ProductNotFoundException exception) {
+
+        return ResponseEntity
+                .status(HttpStatus.NOT_FOUND)
+                .body(ErrorMessage.builder()
+                        .message(exception.getMessage())
+                        .status(HttpStatus.NOT_FOUND.value())
+                        .build());
+    }
+
+    @Override
+    public ResponseEntity<ErrorMessage> updateProductException(UpdateProductException exception) {
+
+        return ResponseEntity
+                .status(HttpStatus.BAD_REQUEST)
+                .body(ErrorMessage.builder()
+                        .message(exception.getMessage())
+                        .status(HttpStatus.BAD_REQUEST.value())
+                        .build());
+    }
+
+    @Override
+    public ResponseEntity<ErrorMessage> notValidException(MethodArgumentNotValidException exception) {
+
+        return ResponseEntity
+                .status(HttpStatus.BAD_REQUEST)
+                .body(ErrorMessage.builder()
+                        .message(exception.getMessage())
+                        .status(HttpStatus.BAD_REQUEST.value())
+                        .build());
+    }
+
+    @Override
+    public ResponseEntity<ErrorMessage> notDataIntegrity(DataIntegrityViolationException exception) {
+
+        return ResponseEntity
+                .status(HttpStatus.CONFLICT)
+                .body(ErrorMessage.builder()
+                        .message(exception.getMessage())
+                        .status(HttpStatus.CONFLICT.value())
+                        .build());
+    }
+}
