--- src/main/java/online/store/controllers/impl/ExceptionApiControllerImpl.java
@@ -1,60 +1,0 @@
-package online.store.controllers.impl;
-
-import online.store.controllers.ExceptionApiController;
-import online.store.exceptions.ErrorMessage;
-import online.store.exceptions.ProductNotFoundException;
-import online.store.exceptions.UpdateProductException;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-
-@ControllerAdvice
-public class ExceptionApiControllerImpl implements ExceptionApiController {
-
-
-    @Override
-    public ResponseEntity<ErrorMessage> notFoundException(ProductNotFoundException exception) {
-
-        return ResponseEntity
-                .status(HttpStatus.NOT_FOUND)
-                .body(ErrorMessage.builder()
-                        .message(exception.getMessage())
-                        .status(HttpStatus.NOT_FOUND.value())
-                        .build());
-    }
-
-    @Override
-    public ResponseEntity<ErrorMessage> updateProductException(UpdateProductException exception) {
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(ErrorMessage.builder()
-                        .message(exception.getMessage())
-                        .status(HttpStatus.BAD_REQUEST.value())
-                        .build());
-    }
-
-    @Override
-    public ResponseEntity<ErrorMessage> notValidException(MethodArgumentNotValidException exception) {
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(ErrorMessage.builder()
-                        .message(exception.getMessage())
-                        .status(HttpStatus.BAD_REQUEST.value())
-                        .build());
-    }
-
-    @Override
-    public ResponseEntity<ErrorMessage> notDataIntegrity(DataIntegrityViolationException exception) {
-
-        return ResponseEntity
-                .status(HttpStatus.CONFLICT)
-                .body(ErrorMessage.builder()
-                        .message(exception.getMessage())
-                        .status(HttpStatus.CONFLICT.value())
-                        .build());
-    }
-}
