+++ src/main/java/online/store/entity/Product.java
@@ -41,14 +41,6 @@ public class Product extends BaseEntity {
     private LocalDateTime lastQuantityUpdatedAt;
 
     @Version
+    @Column(name = "version")
     private Long version;
-
-    public void setQuantity(Integer quantity) {
-        if ((this.quantity == null && quantity != null) ||
-                (this.quantity != null && !this.quantity.equals(quantity))) {
-            this.lastQuantityUpdatedAt = LocalDateTime.now();
-        }
-        this.quantity = quantity;
-    }
-
 }
