+++ src/main/java/online/store/exception/ArticleAlreadyExistsException.java
@@ -1,0 +1,16 @@
+package online.store.exception;
+
+public class ArticleAlreadyExistsException extends RuntimeException {
+
+  public ArticleAlreadyExistsException() {
+    super();
+  }
+
+  public ArticleAlreadyExistsException(String message) {
+    super(message);
+  }
+
+  public ArticleAlreadyExistsException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
