+++ src/main/java/online/store/exception/ErrorMessage.java
@@ -1,0 +1,19 @@
+package online.store.exception;
+
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class ErrorMessage {
+
+    private String message;
+    private Integer status;
+
+    @Builder.Default
+    private LocalDateTime time = LocalDateTime.now();
+
+}
