+++ src/main/java/online/store/exception/ProductNotFoundException.java
@@ -1,0 +1,16 @@
+package online.store.exception;
+
+public class ProductNotFoundException extends RuntimeException {
+
+    public ProductNotFoundException() {
+        super();
+    }
+
+    public ProductNotFoundException(String message) {
+        super(message);
+    }
+
+    public ProductNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
