+++ src/main/java/online/store/exception/UpdateProductException.java
@@ -1,0 +1,7 @@
+package online.store.exception;
+
+public class UpdateProductException extends RuntimeException {
+    public UpdateProductException(String message) {
+        super(message);
+    }
+}
