--- src/main/java/online/store/exceptions/ProductNotFoundException.java
@@ -1,16 +1,0 @@
-package online.store.exceptions;
-
-public class ProductNotFoundException extends RuntimeException {
-
-    public ProductNotFoundException() {
-        super();
-    }
-
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-
-    public ProductNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
