--- src/main/java/online/store/exceptions/UpdateProductException.java
@@ -1,7 +1,0 @@
-package online.store.exceptions;
-
-public class UpdateProductException extends RuntimeException {
-    public UpdateProductException(String message) {
-        super(message);
-    }
-}
