+++ src/main/java/online/store/interfaces/CalculateTransactionTime.java
@@ -1,0 +1,11 @@
+package online.store.interfaces;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface CalculateTransactionTime {
+}
