+++ src/main/resources/application-prod.yaml
@@ -1,12 +1,40 @@
 spring:
+  datasource:
+    driver-class-name: ${DB_DRIVER}
+    username: ${DB_USERNAME}
+    password: ${DB_PASSWORD}
+    url: ${DB_URL}
+
+  jpa:
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
+    show-sql: true
+    hibernate:
+      ddl-auto: none
+  liquibase:
+    change-log: classpath:/db/changelog/db.changelog-master.xml
+    enabled: true
+
   config:
     activate:
       on-profile: prod
   scheduling:
     basic.enabled: false
     optimized.enabled: true
+    fixed.delay: 50000
+    batch.size: 100
 
-fixed:
-  delay: 5000
-batch:
-  size: 100
+
+report:
+  name:
+    report.csv
+
+price:
+  update:
+    multiplier: 1.1
+
+
+springdoc:
+  swagger-ui:
+    path: /swagger-ui.html
