--- src/main/resources/application.yaml
@@ -1,26 +1,0 @@
-spring:
-
-  profiles:
-    active: prod
-
-  datasource:
-    driver-class-name: ${DB_DRIVER}
-    username: ${DB_USERNAME}
-    password: ${DB_PASSWORD}
-    url: ${DB_URL}
-
-  jpa:
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-    show-sql: true
-    hibernate:
-      ddl-auto: none
-  liquibase:
-    change-log: classpath:/db/changelog/db.changelog-master.xml
-    enabled: true
-
-
-springdoc:
-  swagger-ui:
-    path: /swagger-ui.html
